requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
fsspec==0.8.5
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.0.8
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
terminado==0.9.1
test-tube==0.7.5
threadpoolctl==2.1.0
torch==1.7.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.8.1
tornado==6.1
tqdm==4.54.1
tsne==0.3.1
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
